+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-## v26.1.2.24
-* Fixed DirectArray sync
+## v26.1.2.25
+* Fixed node preview rebuilt
+* Fixed editor split window restore